requirement.txt
accelerate==1.7.0
altair==5.5.0
anyio==4.9.0
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
bleach==6.2.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==5.5.2
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
docx2pdf==0.1.8
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fqdn==1.5.1
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gyp==0.1
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.37.0
ipython_genutils==0.2.0
isoduration==20.11.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==5.4.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
more-itertools==8.10.0
mpmath==1.3.0
narwhals==1.42.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydeck==0.9.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-json-logger==3.3.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
safetensors==0.5.3
SecretStorage==3.3.1
Send2Trash==1.8.3
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
ssh-import-id==5.11
stack-data==0.6.3
streamlit==1.45.1
sympy==1.13.1
systemd-python==234
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchinfo==1.8.0
torchvision==0.21.0+cu124
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.2.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xdg==5
xkit==0.0.0
zipp==1.0.0
